requirements.txt
fastapi
uvicorn[standard]
ipython
gunicorn
redis
jsonschema
pyjwt==1.7.1
cryptography
requests
tapisservice
pytest==5.1.2